+++ webpack.config.js
@@ -15,7 +15,14 @@ module.exports = {
   },
   module: {
     loaders: [
-      { test: /(src)(.+)\.js$/, exclude: /node_modules/, loader: 'babel-loader?blacklist=useStrict' }
+      {
+        test: /(src)(.+)\.js$/,
+        exclude: /(node_modules|bower_components)/,
+        loader: 'babel',
+        query: {
+          presets: ['es2015']
+        }
+      }
     ]
   }
 };